src/utils/send_telegram_update.py
"""
Send validation progress message to Telegram
"""
import asyncio
from telegram import Bot

async def send_validation_message():
    bot = Bot(token="YOUR_BOT_TOKEN_HERE")  # Replace with actual token from .env
    chat_id = "YOUR_CHAT_ID"  # Replace with your Telegram user ID
    
    message = """
🔍 COMPREHENSIVE VALIDATION IN PROGRESS

I'm currently processing all 8 sample invoices you provided to validate:

1. GST rate detection (9%, 18%, etc.)
2. Column alignment in Invoice_Header
3. Line items accuracy
4. Customer_Master auto-population  
5. HSN_Master auto-population

Expected completion: 5-10 minutes

I'll send you a detailed report once complete with:
- ✅ What's working correctly
- ⚠️ Any issues found
- 🔧 Fixes applied

You can also test live by sending any invoice to the bot now!
    """
    
    await bot.send_message(chat_id=chat_id, text=message)

if __name__ == "__main__":
    asyncio.run(send_validation_message())
