requirements.txt
# Core dependencies
python-telegram-bot>=21.0
google-generativeai>=0.8.0
gspread>=5.12.0
oauth2client==4.1.3
python-dotenv>=1.0.0
Pillow>=10.1.0
aiohttp>=3.9.1

# Epic 2: Order Upload dependencies
openpyxl>=3.1.0    # Excel file reading for pricing sheet
reportlab>=4.0.0   # PDF generation for clean invoices

# Cloud deployment support (google-generativeai includes google-auth)
